separate PnL download button and report
