+++ src/components/ProfitAndLoss/ProfitAndLoss.tsx
@@ -8,7 +8,9 @@ import { ProfitAndLossChart } from '../ProfitAndLossChart'
 import { ProfitAndLossCompareOptions } from '../ProfitAndLossCompareOptions'
 import { ProfitAndLossDatePicker } from '../ProfitAndLossDatePicker'
 import { ProfitAndLossDetailedCharts } from '../ProfitAndLossDetailedCharts'
+import { ProfitAndLossDownloadButton } from '../ProfitAndLossDownloadButton'
 import { ProfitAndLossHeader } from '../ProfitAndLossHeader'
+import { ProfitAndLossReport } from '../ProfitAndLossReport'
 import { ProfitAndLossSummaries } from '../ProfitAndLossSummaries'
 import { ProfitAndLossTable } from '../ProfitAndLossTable'
 import { endOfMonth, startOfMonth } from 'date-fns'
@@ -79,4 +81,6 @@ ProfitAndLoss.Summaries = ProfitAndLossSummaries
 ProfitAndLoss.Table = ProfitAndLossTable
 ProfitAndLoss.DetailedCharts = ProfitAndLossDetailedCharts
 ProfitAndLoss.Header = ProfitAndLossHeader
+ProfitAndLoss.Report = ProfitAndLossReport
+ProfitAndLoss.DownloadButton = ProfitAndLossDownloadButton
 export { ProfitAndLoss }
