+++ src/components/ProfitAndLossDownloadButton/index.tsx
@@ -1,0 +1,4 @@
+export {
+  ProfitAndLossDownloadButton,
+  PnLDownloadButtonStringOverrides,
+} from './ProfitAndLossDownloadButton'
