+++ src/views/Reports/Reports.tsx
@@ -1,33 +1,26 @@
-import React, { RefObject, useContext, useRef, useState } from 'react'
-import { Layer } from '../../api/layer'
+import React, { RefObject, useState } from 'react'
 import { BalanceSheet } from '../../components/BalanceSheet'
 import { BalanceSheetStringOverrides } from '../../components/BalanceSheet/BalanceSheet'
-import { Button, ButtonVariant, RetryButton } from '../../components/Button'
-import { DownloadButton as DownloadButtonComponent } from '../../components/Button'
 import { Container } from '../../components/Container'
 import { DateRangeDatePickerModes } from '../../components/DatePicker/DatePicker'
-import { Header, HeaderCol, HeaderRow } from '../../components/Header'
-import { Panel } from '../../components/Panel'
 import { ProfitAndLoss } from '../../components/ProfitAndLoss'
 import { ProfitAndLossCompareOptionsProps } from '../../components/ProfitAndLossCompareOptions/ProfitAndLossCompareOptions'
 import { ProfitAndLossDetailedChartsStringOverrides } from '../../components/ProfitAndLossDetailedCharts/ProfitAndLossDetailedCharts'
+import { PnLDownloadButtonStringOverrides } from '../../components/ProfitAndLossDownloadButton'
 import { ProfitAndLossTableStringOverrides } from '../../components/ProfitAndLossTable'
 import { StatementOfCashFlow } from '../../components/StatementOfCashFlow'
 import { StatementOfCashFlowStringOverrides } from '../../components/StatementOfCashFlow/StatementOfCashFlow'
 import { Toggle } from '../../components/Toggle'
 import { View } from '../../components/View'
-import { BREAKPOINTS } from '../../config/general'
-import { useLayerContext } from '../../contexts/LayerContext'
-import { useElementSize } from '../../hooks/useElementSize'
 import { useElementViewSize } from '../../hooks/useElementViewSize'
 import { MoneyFormat } from '../../types'
 import { View as ViewType } from '../../types/general'
 
 type ViewBreakpoint = ViewType | undefined
 
-interface ReportsStringOverrides {
+export interface ReportsStringOverrides {
   title?: string
-  downloadButton?: DownloadButtonStringOverrides
+  downloadButton?: PnLDownloadButtonStringOverrides
   profitAndLoss?: {
     detailedCharts?: ProfitAndLossDetailedChartsStringOverrides
     table?: ProfitAndLossTableStringOverrides
@@ -66,70 +59,6 @@ export interface ReportsPanelProps {
     datePickerMode?: DateRangeDatePickerModes
   }
   view: ViewBreakpoint
-}
-
-interface DownloadButtonStringOverrides {
-  downloadButtonText?: string
-  retryButtonText?: string
-}
-
-interface DownloadButtonProps {
-  stringOverrides?: DownloadButtonStringOverrides
-  moneyFormat?: MoneyFormat
-  view: ViewBreakpoint
-}
-
-const DownloadButton = ({
-  stringOverrides,
-  moneyFormat,
-  view,
-}: DownloadButtonProps) => {
-  const { dateRange } = useContext(ProfitAndLoss.Context)
-  const { auth, businessId, apiUrl } = useLayerContext()
-  const [requestFailed, setRequestFailed] = useState(false)
-  const [isDownloading, setIsDownloading] = useState(false)
-
-  const handleClick = async () => {
-    setIsDownloading(true)
-    const month = (dateRange.startDate.getMonth() + 1).toString()
-    const year = dateRange.startDate.getFullYear().toString()
-    const getProfitAndLossCsv = Layer.getProfitAndLossCsv(
-      apiUrl,
-      auth.access_token,
-      {
-        params: {
-          businessId: businessId,
-          year: year,
-          month: month,
-          moneyFormat: moneyFormat,
-        },
-      },
-    )
-    try {
-      const result = await getProfitAndLossCsv()
-      if (result?.data?.presignedUrl) {
-        window.location.href = result.data.presignedUrl
-        setRequestFailed(false)
-      } else {
-        setRequestFailed(true)
-      }
-    } catch (e) {
-      setRequestFailed(true)
-    } finally {
-      setIsDownloading(false)
-    }
-  }
-
-  return (
-    <DownloadButtonComponent
-      iconOnly={view === 'mobile'}
-      onClick={handleClick}
-      isDownloading={isDownloading}
-      requestFailed={requestFailed}
-      text={stringOverrides?.downloadButtonText || 'Download'}
-      retryText={stringOverrides?.retryButtonText || 'Retry'}
-    />
-  )
 }
 
 const getOptions = (enabledReports: ReportType[]) => {
@@ -218,69 +147,17 @@ const ReportsPanel = ({
   statementOfCashFlowConfig,
   view,
 }: ReportsPanelProps) => {
-  const { sidebarScope } = useContext(ProfitAndLoss.Context)
   return (
     <>
       {openReport === 'profitAndLoss' && (
-        <View
-          type='panel'
-          header={
-            <Header>
-              <HeaderRow>
-                <HeaderCol>
-                  <>
-                    <ProfitAndLoss.DatePicker
-                      datePickerMode={profitAndLossConfig?.datePickerMode}
-                    />
-                    {comparisonConfig && view === 'desktop' ? (
-                      <ProfitAndLoss.CompareOptions
-                        tagComparisonOptions={
-                          comparisonConfig.tagComparisonOptions
-                        }
-                        defaultTagFilter={comparisonConfig.defaultTagFilter}
-                      />
-                    ) : null}
-                  </>
-                </HeaderCol>
-                <HeaderCol>
-                  <DownloadButton
-                    stringOverrides={stringOverrides?.downloadButton}
-                    moneyFormat={profitAndLossConfig?.csvMoneyFormat}
-                    view={view}
-                  />
-                </HeaderCol>
-              </HeaderRow>
-              {comparisonConfig && view !== 'desktop' ? (
-                <HeaderRow>
-                  <HeaderCol>
-                    <ProfitAndLoss.CompareOptions
-                      tagComparisonOptions={
-                        comparisonConfig.tagComparisonOptions
-                      }
-                      defaultTagFilter={comparisonConfig.defaultTagFilter}
-                    />
-                  </HeaderCol>
-                </HeaderRow>
-              ) : null}
-            </Header>
-          }
-        >
-          <Panel
-            sidebar={
-              <ProfitAndLoss.DetailedCharts
-                showDatePicker={false}
-                stringOverrides={stringOverrides?.profitAndLoss?.detailedCharts}
-              />
-            }
-            sidebarIsOpen={Boolean(sidebarScope)}
-            parentRef={containerRef}
-          >
-            <ProfitAndLoss.Table
-              asContainer={false}
-              stringOverrides={stringOverrides?.profitAndLoss?.table}
-            />
-          </Panel>
-        </View>
+        <ProfitAndLoss.Report
+          stringOverrides={stringOverrides}
+          comparisonConfig={comparisonConfig}
+          datePickerMode={profitAndLossConfig?.datePickerMode}
+          csvMoneyFormat={profitAndLossConfig?.csvMoneyFormat}
+          parentRef={containerRef}
+          view={view}
+        />
       )}
       {openReport === 'balanceSheet' && (
         <BalanceSheet stringOverrides={stringOverrides?.balanceSheet} />
